fix: deduplicate utilities across report, campaign, and triage modules
Move load_json_file, parse_timestamp, and format_duration to utils.py
and import from there, eliminating 3 identical copies of load_json_file,
2 copies of parse_timestamp, and 2 copies of format_duration. Also
deduplicate TMP_DIR constant (keep in corpus_manager.py, import in
orchestrator.py). Update test imports accordingly.
