+++ CHANGELOG.md
@@ -81,6 +81,7 @@ All notable changes to this project should be documented in this file.
 
 ### Enhanced
 
+- Deduplicated `load_json_file`, `parse_timestamp`, and `format_duration` utilities from `report.py`, `campaign.py`, and `triage.py` into `lafleur/utils.py`, and deduplicated `TMP_DIR` constant from `orchestrator.py` into `corpus_manager.py`, by @devdanzin.
 - `SniperMutator` with two new attack vectors: `executor_assassinate` (uses `_testinternalcapi.invalidate_executors()` to rip JIT executors out from under active traces, targeting GH-143604) and `globals_detach` (uses `types.FunctionType(func.__code__, new_globals)` to execute JIT-compiled code with detached globals, targeting GH-138378), for both helper functions and builtins, by @devdanzin.
 - `GlobalOptimizationInvalidator` with two new attack vectors: `namespace_swap` (executes JIT-warmed code via `types.FunctionType` with alternate globals containing wrong types, targeting GH-138378) and `globals_dict_mutate` (directly mutates `func.__globals__` in-place after JIT warmup to corrupt cached dict pointers), complementing the existing `evil_global_swap` strategy, by @devdanzin.
 - `UnpackingChaosMutator` with two new `_JitChaosIterator` modes: `backing_store_mutate` (actively modifies the backing list via `clear()`/`extend()`/`insert()`/`pop()` mid-iteration, targeting GH-143123) and `exception_storm` (conditionally raises `StopIteration`/`TypeError`/`ValueError` during iteration), by @devdanzin.
